add support auto-serde in Account and KeyedAccount (move Errors to help)
