+++ programs/config_api/src/config_processor.rs
@@ -2,7 +2,7 @@
 
 use log::*;
 use solana_sdk::account::KeyedAccount;
-use solana_sdk::instruction::InstructionError;
+use solana_sdk::instruction_error::InstructionError;
 use solana_sdk::pubkey::Pubkey;
 
 pub fn process_instruction(
