+++ programs/exchange_api/src/exchange_processor.rs
@@ -5,7 +5,7 @@ use crate::exchange_state::*;
 use crate::id;
 use log::*;
 use solana_sdk::account::KeyedAccount;
-use solana_sdk::instruction::InstructionError;
+use solana_sdk::instruction_error::InstructionError;
 use solana_sdk::pubkey::Pubkey;
 use std::cmp;
