+++ sdk/src/lib.rs
@@ -3,6 +3,7 @@ pub mod bpf_loader;
 pub mod genesis_block;
 pub mod hash;
 pub mod instruction;
+pub mod instruction_error;
 pub mod loader_instruction;
 pub mod message;
 pub mod native_loader;
@@ -12,6 +13,7 @@ pub mod pubkey;
 pub mod rpc_port;
 pub mod short_vec;
 pub mod signature;
+pub mod system_error;
 pub mod system_instruction;
 pub mod system_program;
 pub mod system_transaction;
