+++ sdk/src/system_error.rs
@@ -1,0 +1,6 @@
+#[derive(Serialize, Debug, Clone, PartialEq)]
+pub enum SystemError {
+    AccountAlreadyInUse,
+    ResultWithNegativeLamports,
+    SourceNotSystemAccount,
+}
